requirements.txt
streamlit==1.31.0
pandas==2.2.0
fpdf2==2.7.8